Move auth-controller to components/core
