+++ src/components/lineup-view-game-detail.ts
@@ -7,13 +7,13 @@ import { updateMetadata } from 'pwa-helpers/metadata.js';
 import { RootState } from '../app/store.js';
 import { GameDetail, GameStatus } from '../models/game.js';
 import { getGame, getGameSliceConfigurator, selectGameById } from '../slices/game/index.js';
+import { SignedInAuthController } from './core/auth-controller.js';
 import { ConnectStoreMixin } from './core/connect-mixin.js';
 import './lineup-game-complete.js';
 import './lineup-game-live.js';
 import './lineup-game-setup.js';
 import { AuthorizedViewElement } from './page-view-element.js';
 import { SharedStyles } from './shared-styles.js';
-import { SignedInAuthController } from './util/auth-controller.js';
 
 @customElement('lineup-view-game-detail')
 export class LineupViewGameDetail extends ConnectStoreMixin(AuthorizedViewElement) {
