+++ src/components/lineup-view-games.ts
@@ -7,6 +7,7 @@ import { RootState } from '../app/store.js';
 import { Game, Games } from '../models/game.js';
 import { selectCurrentTeam } from '../slices/app/index.js';
 import { addNewGame, getGameSliceConfigurator, getGames } from '../slices/game/index.js';
+import { SignedInAuthController } from './core/auth-controller.js';
 import { ConnectStoreMixin } from './core/connect-mixin.js';
 import { logger } from '../util/logger.js';
 import './lineup-game-create.js';
@@ -14,7 +15,6 @@ import { GameCreatedEvent, LineupGameCreate } from './lineup-game-create.js';
 import './lineup-game-list.js';
 import { AuthorizedViewElement } from './page-view-element.js';
 import { SharedStyles } from './shared-styles.js';
-import { SignedInAuthController } from './util/auth-controller.js';
 
 const debugGames = logger('view-games');
